Basic filtering

+++ backend/src/main/java/com/example/demo/model/Product.java
@@ -3,18 +3,19 @@ package com.example.demo.model;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
 
+import java.math.BigDecimal;
 import java.sql.Timestamp;
 
 @Getter
 public class Product {
     final String name;
     final int stock;
-    final double price;
+    final BigDecimal price;
     final String image_url;
 
     public Product(@JsonProperty("name") String name,
                    @JsonProperty("stock") int stock,
-                   @JsonProperty("price") double price,
+                   @JsonProperty("price") BigDecimal price,
                    @JsonProperty("image_url") String image_url) {
         this.name = name;
         this.stock = stock;
